Make the observation type a float for better granularity

+++ gym_solo/solo_types.py
@@ -1,4 +1,4 @@
 from typing import List
 
 # A state observation
-obs = List[int] 
+obs = List[float]
